+++ common/SpookyV2.h
@@ -71,7 +71,7 @@ public:
     {
         uint64 hash1 = seed, hash2 = seed;
         Hash128(message, length, &hash1, &hash2);
-        return (uint32)hash1;
+        return hash1;
     }
 
     static uint64_t hashSubBuffer(unsigned char* pixmap, size_t startX, size_t startY,
